requirements.txt
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
distlib==0.4.0
distro==1.9.0
einops==0.8.1
filelock==3.20.3
filetype==1.2.0
fsspec==2026.1.0
ftfy==6.3.1
google-auth==2.47.0
google-genai==1.59.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
identify==2.6.16
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
markdown2==2.5.4
markdownify==1.2.2
marker-pdf==1.10.1
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==1.109.1
opencv-python-headless==4.11.0.86
packaging==25.0
pdftext==0.6.3
pillow==10.4.0
platformdirs==4.5.1
pre_commit==4.5.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pypdfium2==4.30.0
python-dotenv==1.2.1
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2024.11.6
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
surya-ocr==0.17.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.6
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.36.1
wcwidth==0.2.14
websockets==15.0.1
